README.md
Use the WEEX referral code 3AND to unlock a bonus up to 30000 USDT and enjoy up to 50% discount on trading fees. Whether you're a beginner or an experienced trader, this referral code gives you access to WEEX’s powerful platform and exclusive rewards. This complete guide explains how to use the WEEX referral program effectively, the benefits of codes like 3AND and TOPTOP, and how to get started today.